Moved the loader in the utils
