+++ debiaiServer/utils/utils.py
@@ -1,11 +1,12 @@
 import time
-import yaml
 import pkg_resources
-from yaml.loader import SafeLoader
 from urllib.parse import urlparse
 
 
 def get_app_version():
+    import yaml
+    from yaml.loader import SafeLoader
+
     # Read the version from the API YAML file
     yaml_path = pkg_resources.resource_filename("debiaiServer", "swagger.yaml")
 
@@ -47,3 +48,51 @@ def is_valid_name(name):
         return False
 
     return True
+
+
+class Loader:
+    from itertools import cycle
+    from shutil import get_terminal_size
+    from threading import Thread
+    from time import sleep
+
+    def __init__(self, desc="Loading...", end="Done!", timeout=0.1):
+        """
+        A loader-like context manager
+
+        Args:
+            desc (str, optional): The loader's description. Defaults to "Loading...".
+            end (str, optional): Final print. Defaults to "Done!".
+            timeout (float, optional): Sleep time between prints. Defaults to 0.1.
+        """
+        self.desc = desc
+        self.end = end
+        self.timeout = timeout
+
+        self._thread = Thread(target=self._animate, daemon=True)
+        self.steps = ["⢿", "⣻", "⣽", "⣾", "⣷", "⣯", "⣟", "⡿"]
+        self.done = False
+
+    def start(self):
+        self._thread.start()
+        return self
+
+    def _animate(self):
+        for c in cycle(self.steps):
+            if self.done:
+                break
+            print(f"\r{self.desc} {c}", flush=True, end="")
+            sleep(self.timeout)
+
+    def __enter__(self):
+        self.start()
+
+    def stop(self):
+        self.done = True
+        cols = get_terminal_size((80, 20)).columns
+        print("\r" + " " * cols, end="", flush=True)
+        print(f"\r{self.end}", flush=True)
+
+    def __exit__(self, exc_type, exc_value, tb):
+        # handle exceptions with those variables ^
+        self.stop()
